+++ src/layers/mod.rs
@@ -1,2 +1,1 @@
-pub mod activation_functions;
 pub mod linear;
